services/orchest-webserver/app/requirements.txt
nbconvert~=6.0.0
IPython
Flask
flask_sqlalchemy
flask_restful
flask-marshmallow
flask-socketio
eventlet
marshmallow-sqlalchemy
Flask-Migrate
requests
APScheduler
nbformat
docker
Werkzeug==0.16.0
python-socketio[client]
-e ../../../lib/python/orchest-internals
